Updated godot-cpp to godot-jolt/godot-cpp@6fbafc4ba3 (#179)

+++ src/jolt_joint_3d.cpp
@@ -5,7 +5,7 @@
 
 namespace {
 
-constexpr int64_t GDJOLT_JOINT_DEFAULT_SOLVER_PRIORITY = 1;
+constexpr int32_t GDJOLT_JOINT_DEFAULT_SOLVER_PRIORITY = 1;
 
 } // namespace
 
@@ -20,11 +20,11 @@ JoltJoint3D::~JoltJoint3D() {
 	}
 }
 
-int64_t JoltJoint3D::get_solver_priority() const {
+int32_t JoltJoint3D::get_solver_priority() const {
 	return GDJOLT_JOINT_DEFAULT_SOLVER_PRIORITY;
 }
 
-void JoltJoint3D::set_solver_priority(int64_t p_priority) {
+void JoltJoint3D::set_solver_priority(int32_t p_priority) {
 	if (p_priority != GDJOLT_JOINT_DEFAULT_SOLVER_PRIORITY) {
 		WARN_PRINT(
 			"Joint solver priority is not supported by Godot Jolt. "
